Add files via upload
This code implements a simple text-based game called "ODD/IF" which is similar to cricket. The game involves two players: the user and the computer. The main steps are:

#Game Rules Explanation: The rules of the game are printed at the beginning.

1. User Input: The user chooses between "ODD" or "IF" (even) and enters a number between 1 and 10.
2. Toss: A toss is conducted by summing the user's and computer's numbers. The outcome (odd or even) determines who wins the toss.
3. Batting/Bowling: Based on the toss result, the user or computer chooses to bat or bowl. The game proceeds with the user and computer taking turns to bat and bowl.
4. Scoring: During batting, if the chosen numbers by the user and computer match, the batsman is out. The score is accumulated until the batsman is out or the target is reached.
5. Winning Condition: The game checks the scores to determine the winner.

#Key functions:

choose_number(): Prompts the user to enter a valid number between 1 and 10.
toss(total_num, uc): Determines the toss result and who gets to choose batting or bowling.
user_batting(t=0): Handles the user's batting logic.
user_bowling(t=0): Handles the user's bowling logic.

The game loop continues until the user decides to quit.